pyproject.toml
[project]
name = "yolo-v1"
version = "0.1.0"
description = "Reimplementation of the YOLO v1 object detection model using PyTorch."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["torch>=2.8.0", "torchvision>=0.23.0", "pillow>=11.3.0"]

[dependency-groups]
dev = ["pytest>=8.4.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yolo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short"]
